Modify include behavior to act as reference, not copy
This changes the way `git hooks include` works so that rather than
physically copying the included script to the destination repo's
.githooks directory, it instead creates a git config file in the
.githooks directory that indicates which included script to run.
`git hooks` is then smart enough to locate that script in the
`git-hooks/included` directory and execute it.

This is so that one can reference shared hooks and benefit from updates
to the `git-hooks` project rather than having to manually update them by
copying code over.