Rename alias->command, style updates
Renames `git hooks install-alias` to `git hooks install-command`, since the
desired effect is to install the `git hooks` command. The fact that it's
accomplished via an alias is an incidental implementation detail.

It also changes the default setting to install the command as a global alias
rather than a local one.

Other minor changes:
- Removed the `--examples` option to the `install` command
- Enabled support for symlinking to the git-hooks script
- Add the pre-commit-01-readme script to automatically update README.md