Look for collection subpath in the collection itself first
This let's a collection maintainer specify where to find the hooks directory
if the repo root is not the hooks directory. We still allow the user to override
this, as it supports the case where they may want to use a non-conforming hooks
collection repo where the collection has not provided the subpath.