[ENG-63] Split the prepare-commit-msg/template into two hooks (#6)
Split the `prepare-commit-msg/template` into two hooks: `prepare-commit-msg/cached` and
`prepare-commit-msg/subject`.

The `cached` hook will check for any commit messages stored by a later hook during downstream
validation (during a previous commit attempt). If found, it will display the contents of the
cached message and prompt the user to use it, ignore it, or discard it. Ignoring it will leave
the cached message in place for later attempts, but discarding it will remove it completely.

The `subject` hook enables one to provide the commit message on the command-line with the
`git commit -m "<message>"` option and still get the benefits of using a commit message template.
One usually provides a template with the `commit.template` git config value. If the template
contains the `$SUBJECT` placeholder, the provided message will be substituted there. If no message
is provided at the command-line, this hook simply removes the `$SUBJECT` placeholder prior to the
message being presented to the user for editing.

Prior to these changes, it was assumed that the commit template file could be found at
`<repo-root>/.gitmessage`. However, we now rely on the user leveraging existing git functionality
and providing the commit message template by conventional means. This means that one may also
specify a global commit message template, say in their home directory, and have it applied to all git
repos on their machine. This, combined with the global git hooks, make for a consistent experience
all around.

We are also renaming `post-commit/template-cleanup` to `post-commit/cached-cleanup` to match
the new `prepare-commit-msg/cached-cleanup`.

Furthermore, `prepare-commit-msg/template` and `post-commit/template-cleanup` now display
deprecation warnings (but otherwise remain functional).

[ENG-63](https://fivestars.atlassian.net/browse/ENG-63)