[ENG-69] Create Sentry links from Jira notations in commit messages
This adds the `commit-msg/append-sentry-issues` hook to automatically update the
commit message with `Fixes SENTRY-1234` entries for any Jira issues found in the
commit message that were created by Sentry. This is important for Sentry integrations
as it allows Sentry to automatically close an issue if it's marked as part of a release.

[ENG-69](https://fivestars.atlassian.net/browse/ENG-69)