[ENG-80] Prepopulate a commit message with the Jira ticket info
We now pull the summary and description info from Jira since we know the Jira
ticket number from the branch name.

If the git message template has certain placeholders like `$JIRA_SUMMARY` or
`$JIRA_DESC`, they will be replaced with the appropriate content from the
ticket.

This only occurs if the parent commit does not reference the current Jira
ticket. This is to prevent subsequent commits on a feature branch from being
prepopulated with the ticket info again (as it would be redundant).

[ENG-80](https://fivestars.atlassian.net/browse/ENG-80)