[ENG-81] Populate commit message with Sentry issue info
This adds `included/hooks/commit-msg/add-sentry-issues`. It will search any Jira
tickets referenced in the commit message for a Sentry issue link. If found, it
adds a `Fixes <sentry issue>` line to the commit message. The `Resolved Sentry
Issue` section will be cleared and recreated any time the commit message is
updated. This is to ensure that if one adds a Jira ticket to the message, it
will be included in the Sentry section as well.

[ENG-81](https://fivestars.atlassian.net/browse/ENG-81)