[ENG-81] Populate commit message with Sentry issue info (#12)
This adds `included/hooks/commit-msg/add-sentry-issues`. It will search any Jira
tickets referenced in the commit message for a Sentry issue link. If found, it
adds a `Fixes <sentry issue>` line to the commit message in a `Resolved Sentry Issue`
section at the end of the commit section. This section will be cleared and recreated
any time the commit message is updated. This is to ensure that if one adds a Jira
ticket to the message, it will be included in the Sentry section as well.

Because this hook is a `commit-msg` hook, it will run after you finish editing your
commit message. This could collide with commit message verification hooks like
`format` and `jira-format`. Since this new hook can insert lines that are quite long
in markdown syntax (but short when rendered), we also provide a way to bypass
length checks on these potential lines. One can set the `hooks.format.long-line-regex`
or `hooks.jira-format.long-line-regex` value (according to which hook you're using
to validate commit messages). The regex will be applied with the `-E` flag, meaning
one can use "extended" regular expressions.

[ENG-81](https://fivestars.atlassian.net/browse/ENG-81)