[ENG-83] Add output to prepare-commit-msg hooks when they are skipped
`prepare-commit-msg` hooks get some input values that help indicate what
operation is occurring that requires a commit message to be created or edited.
This uses these values to determine when no work should take place for the hook
and have the hook report that it is being skipped and why.

https://git-scm.com/docs/githooks#_prepare_commit_msg

[ENG-83](https://fivestars.atlassian.net/browse/ENG-83)