[ENG-85] Allow non-Jira-based named branches for unticketed work
This supports the use case where one might wish to create some feature branches
that are not associated with a Jira ticket. It allows for one to specify a
magic-string prefix, which defaults to `local`, to indicate that a branch
should not be subjected to the Jira-based hooks' checks and manipulations.

If you wish to use something other than `local`, set the
`git-hooks.non-ticketed-prefix` git config value.

[ENG-85](https://fivestars.atlassian.net/browse/ENG-85)