[ENG-88] Check for updates periodically (#19)
It now checks once a week for any new release tags. If found, it will prompt the
user to decide whether to update at that time, then performs the update if so
instructed. It will checkout the latest release tag (v#.#.#) ahead of the
current HEAD.

If the update check fails (due to say network connectivity issues), it waits 1
day before attempting again.

This also adds an `update` command to allow one to manually check for updates.

[ENG-88](https://fivestars.atlassian.net/browse/ENG-88)