[ENG-92] Fix implicit sync-collection
There was a subtle bug in this functionality that prevented the collection repos
from updating during the weekly check. We no longer wrap the `$force` parameter
in double-quotes when calling `git_hooks__update_if_needed`, as it will be
interpreted as a blank string for `$1` rather than as non-present when `$force`
not set.

[ENG-92](https://fivestars.atlassian.net/browse/ENG-92)