[ENG-93] Enable "global sets" of global hooks (#25)
It's an expected use case that one may develop in several projects in different
roles as a developer. For example, one may work in a set of repos owned by their
employer, and another set of repos that are open-source side projects. The
job-related repos may require a different set of global hooks than the
open-source ones.

This adds a way to configure sets of global hooks such that a repository can be
associated with at most one of them.

It creates the new `~/.githooks/.global-sets/` directory to hold each global set
of hooks (in subdirectories). A global set consists of a set of reference hook
scripts. They will reference a shared set of collections.

For now, to maintain backwards compatibility, the default anonymous global set
will existing in the `~/.githooks/` directory, while other named ones will exist
in named subdirectories of the `~/.githooks/.global-sets/` directory.

This adds two new `git hooks` commands: `use-globals` and `disable-globals`

`use-globals` accepts an optional name to indicate which global set to use for
the current repository. If a name is not provided, it's assumed that you wish to
use the anonymous default global set.

`disable-globals` will instruct the repository to not use any global hooks. Use
`use-globals` to undo this operation.

[ENG-93](https://fivestars.atlassian.net/browse/ENG-93)