[ENG-113] Only use global_githooks_dir if including a global hook
The code was attempting to use the global_githooks_dir, even when the `-g` flag was not specified.
This resulted in an undefined variable error when one tried to include a hook in a local
repository's hooks (and not a global set).

This changes it so that the installation code uses `$githooks_dir` and ensures that it is set
correctly according to the `-g` flag's setting.

[ENG-113](https://fivestars.atlassian.net/browse/ENG-113)