[ENG-116] Improve the way we translate Jira markup to commit message markdown
We now automatically translate Jira links and `{{code}}` blocks to markdown syntax when preparing a
commit message in the `prepare-commit-msg/jira` hook.

Also, we no longer fold any lines to a fixed length if the issue was created by an outside
integration and wasn't manually typed in by a developer.

[ENG-116](https://fivestars.atlassian.net/browse/ENG-116)